fix: fix StatTile and move pages to src
