+++ src/components/StatTile.tsx
@@ -1,3 +1,5 @@
+import { fr } from "@codegouvfr/react-dsfr";
+
 export type StatTileProps = {
   title: string;
   stats: string | number;
@@ -6,13 +8,13 @@ export type StatTileProps = {
 
 export const StatTile = (props: StatTileProps): JSX.Element => {
   return (
-    <div className="fr-col-12 fr-col-md-3">
-      <div className="fr-card fr-centered fr-card--no-arrow">
-        <div className="fr-card__body">
-          <strong className="fr-display-xs">{props.stats}</strong>
-          <h2 className="fr-card__title fr-mb-4w">{props.title}</h2>
+    <div className={fr.cx("fr-col-12", "fr-col-md-4")}>
+      <div className={fr.cx("fr-card", "fr-card--no-arrow")}>
+        <div className={fr.cx("fr-card__body")}>
+          <strong className={fr.cx("fr-display--md")}>{props.stats}</strong>
+          <h2 className={fr.cx("fr-card__title", "fr-mb-4w")}>{props.title}</h2>
           {props.description && (
-            <div className="fr-card__desc">
+            <div className={fr.cx("fr-card__desc")}>
               <p>{props.description}</p>
             </div>
           )}
